+++ src/main/java/net/meander/subtlyd/mixin/client/gui/screens/FaviconTextureMixin.java
@@ -22,7 +22,7 @@ public class FaviconTextureMixin {
     @Shadow @Final private TextureManager textureManager;
     @Shadow private void checkOpen() {}
     @Shadow public void clear() {}
-    private static final boolean canChangeUi = OptionsSD.EXPERIMENTAL_GUI.get();
+    private static final boolean canChangeUi = OptionsSD.gui().get();
 
     /**
      * Saves a thumbnail of size 455x256 pixels