+++ src/main/java/net/meander/subtlyd/mixin/client/gui/screens/worldselection/CreateWorldScreenMixin.java
@@ -49,7 +49,7 @@ public abstract class CreateWorldScreenMixin extends Screen {
     @Shadow @Final @Mutable private final HeaderAndFooterLayout layout = new HeaderAndFooterLayout(this);
     @Shadow @Final private WorldCreationUiState uiState;
     @Shadow private MenuTabBar tabNavigationBar;
-    private static final boolean canChangeUi = OptionsSD.EXPERIMENTAL_GUI.get();
+    private static final boolean canChangeUi = OptionsSD.gui().get();
 
     protected CreateWorldScreenMixin(Component title) {
         super(title);
@@ -125,7 +125,7 @@ public abstract class CreateWorldScreenMixin extends Screen {
         private static final Component ALLOW_COMMANDS_INFO = Component.translatable("selectWorld.allowCommands.info");
         private static final Component HARDCORE = Component.translatable("selectWorld.gameMode.hardcore");
         private static final Component HARDCORE_INFO = Component.translatable("selectWorld.gameMode.hardcore.info");
-        private static final boolean canChangeUi = OptionsSD.EXPERIMENTAL_GUI.get();
+        private static final boolean canChangeUi = OptionsSD.gui().get();
 
         public GameTabMixin(EditBox nameEdit) {
             super(TITLE);