+++ src/main/java/net/meander/subtlyd/mixin/client/options/OptionsMixin.java
@@ -1,5 +1,7 @@
 package net.meander.subtlyd.mixin.client.options;
 
+import net.fabricmc.api.EnvType;
+import net.fabricmc.api.Environment;
 import net.meander.subtlyd.client.OptionsSD;
 import net.minecraft.client.Options;
 import org.spongepowered.asm.mixin.Mixin;
@@ -7,15 +9,14 @@ import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
+@Environment(EnvType.CLIENT)
 @Mixin(Options.class)
 public class OptionsMixin {
-    /**
-     * Saves the new custom options
-     */
     @Inject(method = "processOptions", at = @At("HEAD"))
     private void saveOptions(Options.FieldAccess access, CallbackInfo ci) {
-        access.process("camera_shake", OptionsSD.CAMERA_SHAKE);
-        access.process("experimental_gui", OptionsSD.EXPERIMENTAL_GUI);
-        access.process("shield_crouch", OptionsSD.SHIELD_CROUCH);
+        access.process("camera_shake", OptionsSD.cameraShake());
+        access.process("experimental_gui", OptionsSD.gui());
+        access.process("shield_crouch", OptionsSD.shieldCrouch());
+        access.process("shield_animation", OptionsSD.shieldAnimation());
     }
 }