+++ src/main/java/net/meander/subtlyd/mixin/client/renderer/entity/LivingEntityRendererMixin.java
@@ -33,7 +33,7 @@ public class LivingEntityRendererMixin<T extends LivingEntity, S extends LivingE
     @Inject(method = "extractRenderState(Lnet/minecraft/world/entity/LivingEntity;Lnet/minecraft/client/renderer/entity/state/LivingEntityRenderState;F)V",
             at = @At("TAIL"))
     private void determineRenderState(T entity, S state, float partialTicks, CallbackInfo ci) {
-        if (OptionsSD.ADVANCED_ENTITY_ANIMATIONS.get()) {
+        if (OptionsSD.advancedEntityAnimations().get()) {
             if (state instanceof LivingEntityRenderStateAccessor stateAccessor) {
                 if (entity instanceof ClimberAccessor climberAccessor) {
                     extractClimberState(entity, climberAccessor, state, stateAccessor, partialTicks);
@@ -51,7 +51,7 @@ public class LivingEntityRendererMixin<T extends LivingEntity, S extends LivingE
      */
     @Inject(method = "setupRotations", at = @At("TAIL"))
     private  void determineRotations(S state, PoseStack poseStack, float bodyRot, float entityScale, CallbackInfo ci) {
-        if (OptionsSD.ADVANCED_ENTITY_ANIMATIONS.get()) {
+        if (OptionsSD.advancedEntityAnimations().get()) {
             if (state instanceof LivingEntityRenderStateAccessor accessor) {
                 setupClimberRotations(accessor, poseStack, state.boundingBoxHeight);
             }